+++ CHANGES.rst
@@ -1,6 +1,7 @@
 0.2.3
 -----
 - Add tests/default.conf to sdist/pip tarball. #27
+- Do not allow any dropped cells if Tor 0.3.5.1-alpha+ is being used. #25 and #3.
 
 0.2.2
 -----